Add patch verification tests for remove and replace operations
