+++ cmd/patch_test.go
@@ -21,15 +21,12 @@ package cmd
 import (
 	"bufio"
 	"bytes"
-	"fmt"
 	"io"
 	"log"
 	"os"
-	"reflect"
 	"strings"
 	"testing"
 
-	"github.com/CycloneDX/sbom-utility/common"
 	"github.com/CycloneDX/sbom-utility/schema"
 	"github.com/CycloneDX/sbom-utility/utils"
 )
@@ -234,181 +231,6 @@ func retrieveQueryPathFromPatchRecord(recordPath string) (queryPath string, key 
 	} else {
 		queryPath = paths[0]
 		// NOTE: default root (i.e., "")
-	}
-	return
-}
-
-func VerifyPatchedOutputFileResult(t *testing.T, originalTest PatchTestInfo) (err error) {
-	getLogger().Enter()
-	defer getLogger().Exit()
-
-	patchDocument := NewIETFRFC6902PatchDocument(originalTest.PatchFile)
-	if err = patchDocument.UnmarshalRecords(); err != nil {
-		return
-	}
-
-	// If no patch records were found after unmarshal
-	if patchDocument.Records == nil {
-		return
-	}
-
-	// Create a new test info. structure copying in data from the original test
-	queryTestInfo := NewCommonTestInfo()
-	queryTestInfo.InputFile = originalTest.OutputFile
-
-	// Load and Query temporary "patched" output BOM file using the "from" path
-	// Default to "root" (i.e,, "") path if none selected.
-	DEFAULT_PATH_DOC_ROOT := ""
-	request, err := common.NewQueryRequestSelectFromWhere(
-		common.QUERY_TOKEN_WILDCARD, DEFAULT_PATH_DOC_ROOT, "")
-	if err != nil {
-		t.Errorf("%s: %v", ERR_TYPE_UNEXPECTED_ERROR, err)
-		return
-	}
-	getLogger().Tracef("request: %s\n", request.String())
-
-	// Verify each key was removed
-	var pResult interface{}
-	for _, record := range patchDocument.Records {
-		var queryPath, key string
-		queryPath, key, err = retrieveQueryPathFromPatchRecord(record.Path)
-		//fmt.Printf("queryPath: %s, key: %s\n", queryPath, key)
-		if err != nil {
-			t.Errorf("%s: %v", "unable to parse patch record path.", err)
-			return
-		}
-		request.SetRawFromPaths(queryPath)
-
-		// use a buffered query on the temp. output file on the (parent) path
-		pResult, _, err = innerQuery(t, queryTestInfo, request)
-
-		// NOTE: Query typically does NOT support non JSON map or slice
-		// we need to allow float64, bool and string for "patch" validation
-		if err != nil && !ErrorTypesMatch(err, &common.QueryResultInvalidTypeError{}) {
-			t.Errorf("%s: %v", ERR_TYPE_UNEXPECTED_ERROR, err)
-			return
-		}
-
-		// short-circuit if the "from" path dereferenced to a non-existent key
-		if pResult == nil {
-			t.Errorf("empty (nil) found at from clause: %s", request.String())
-			return
-		}
-
-		// verify the "key" was removed from the (parent) JSON map
-		err = VerifyPatched(record, pResult, key)
-		if err != nil {
-			return
-		}
-	}
-
-	return
-}
-
-func VerifyPatched(record IETF6902Record, pResult interface{}, key string) (err error) {
-	getLogger().Enter()
-	defer getLogger().Exit()
-
-	// verify the "key" was removed from the (parent) JSON map
-	if pResult != nil {
-		switch typedResult := pResult.(type) {
-		case map[string]interface{}:
-			// NOTE: this is for "Add" operation only
-			if record.Operation == IETF_RFC6902_OP_ADD {
-				if _, ok := typedResult[key]; !ok {
-					formattedResult, _ := utils.EncodeAnyToDefaultIndentedJSONStr(typedResult)
-					err = getLogger().Errorf("patch failed. Key `%s`, found in: `%s`", key, formattedResult.String())
-					return
-				}
-			}
-		case []interface{}:
-			// NOTE: this is for "Add" operation only
-			if record.Operation == IETF_RFC6902_OP_ADD {
-				if len(typedResult) == 0 {
-					err = getLogger().Errorf("verify failed. Record slice value is empty.")
-					return
-				}
-
-				if record.Value == nil {
-					err = getLogger().Errorf("verify failed. Document slice value is nil.")
-					return
-				}
-				var contains bool
-				contains, err = sliceContainsValue(typedResult, record.Value)
-				if !contains {
-					err = getLogger().Errorf("verify failed. Document value (%v) does not contain expected value (%v).", typedResult, record.Value)
-					return
-				}
-			}
-		case string:
-			if record.Operation == IETF_RFC6902_OP_ADD {
-				if record.Value != typedResult {
-					err = getLogger().Errorf("verify failed. Document value (%v) does not contain expected value (%v).", typedResult, record.Value)
-					return
-				}
-			}
-			return
-		case float64: // NOTE: encoding/json turns int64 to float64
-			if record.Operation == IETF_RFC6902_OP_ADD {
-				if record.Value != typedResult {
-					err = getLogger().Errorf("verify failed. Document value (%v) does not contain expected value (%v).", typedResult, record.Value)
-					return
-				}
-			}
-			return
-		case bool:
-			if record.Operation == IETF_RFC6902_OP_ADD {
-				if record.Value != typedResult {
-					err = getLogger().Errorf("verify failed. Document value (%v) does not contain expected value (%v).", typedResult, record.Value)
-					return
-				}
-			}
-			return
-		default:
-			err = getLogger().Errorf("verify failed. Unexpected JSON type: `%T`", typedResult)
-			return
-		}
-	} else {
-		// TODO: return typed error
-		getLogger().Trace("nil results")
-	}
-	return
-}
-
-func sliceContainsValue(slice []interface{}, value interface{}) (contains bool, err error) {
-	getLogger().Enter()
-	defer getLogger().Exit()
-
-	switch typedValue := value.(type) {
-	case map[string]interface{}:
-		var tempValue map[string]interface{}
-		var ok bool
-		for _, entry := range slice {
-			if tempValue, ok = entry.(map[string]interface{}); !ok {
-				err = fmt.Errorf("type mismatch error. Slice values: %v (%T), value: %v (%T)", entry, entry, tempValue, tempValue)
-				return
-			}
-			if reflect.DeepEqual(tempValue, typedValue) {
-				contains = true
-				return
-			}
-		}
-		return
-	case []interface{}:
-		if reflect.DeepEqual(slice, typedValue) {
-			contains = true
-			return
-		}
-		return
-	case string: // TODO
-		// NOTE: "slices.Contains()" does not work on []interface{}
-		return
-	case bool: // TODO
-		return
-	case float64: // TODO
-		return
-	default:
-		getLogger().Errorf("contains test failed. Unexpected JSON type: `%T`", typedValue)
 	}
 	return
 }
@@ -916,6 +738,7 @@ func TestPatchCdx15Example5ReplaceMetadataTimestamp(t *testing.T) {
 	}
 }
 
+// { "op": "remove", "path": "/metadata/properties/1" }
 func TestPatchCdx15Example6RemoveMetadataProperty(t *testing.T) {
 	ti := NewPatchTestInfo(TEST_PATCH_BOM_1_5_SIMPLEST_BASE, TEST_PATCH_EXAMPLE_REMOVE_METADATA_PROPERTY, nil)
 	ti.OutputFile = ti.CreateTemporaryTestOutputFilename(TEST_PATCH_BOM_1_5_SIMPLEST_BASE)
